channeldb: remove address resolution from channeldb
